# src/main/resources/api.yaml
openapi: 3.0.1
info:
  description: Vaibhav Jain's Employee Management System
  version: 1.0.0
  title: Vaibhav Jain's Employee Management System
  contact:
    name: Vaibhav Jain
  license:
    name: 'Copyright (c) 2023 Vaibhav, Inc. All rights reserved.'
servers:
  - url: '/api/v1'
tags:
  - name: employees-data
    description: 'Activities-data related logical API group'
paths:
  '/api/v1/employees':
    post:
      tags:
        - employees-data
      summary: Create employees
      description : create employees without uuid
      operationId: createEmployee
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeRequest'
        description: Create employee request
        required: false
        x-Name: bre-init
      responses:
        200:
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    get:
      tags:
        - employees-data
      summary: Retrieve all employees data
      description: Get all employees
      operationId: getAllEmployeesData
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeesData'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found

  '/api/v1/employees/{employeeId}':
    get:
      tags:
        - employees-data
      summary: Retrieve employee from employee ID
      description: Get all employees
      operationId: getEmployee
      parameters:
        - $ref: '#/components/parameters/EmployeeIdPathParam'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found


components:
  parameters:
    EmployeeIdPathParam:
      $ref: 'api-params.yaml#/EmployeeIdPathParam'
  schemas:
    GetActivitiesQuery:
      $ref: 'api-defs.yaml#/GetActivitiesQuery'
    EmployeesData:
      $ref: 'api-defs.yaml#/EmployeesData'
    CreateEmployeeRequest:
      $ref: 'api-defs.yaml#/CreateEmployeeRequest'
    Employee:
      $ref: 'api-defs.yaml#/Employee'